BREAKING CHANGE: onboarding models and API moved to spid_cie_oidc.authority
