+++ spid_cie_oidc/authority/urls.py
@@ -17,15 +17,13 @@ from django.conf import settings
 from django.urls import path
 
 from .views import (
-    onboarding_landing, 
-    onboarding_registration, 
-    onboarding_entities
+    fetch, 
+    entity_list
 )
 
 _PREF = getattr(settings, "OIDC_PREFIX", "")
 
 urlpatterns = [
-    path(f"{_PREF}onboarding/landing/", onboarding_landing, name="oidc_onboarding_landing"),
-    path(f"{_PREF}onboarding/registration/", onboarding_registration, name="oidc_onboarding_registration"),
-    path(f"{_PREF}onboarding/entities/", onboarding_entities, name="oidc_onboarding_entities"),
+    path(f"{_PREF}fetch/", fetch, name="oidcfed_fetch"),
+    path(f"{_PREF}list/", entity_list, name="oidcfed_list"),
 ]
